@@ -1,7 +1,7 @@
 package com.example.waveloadingview
 
 import android.graphics.Path
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
 class DemoActivity : AppCompatActivity() {
@@ -9,9 +9,6 @@ class DemoActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_demo)
-
-
-
 
     }
 }
